feat: add migration functionality

+++ library/core/src/commonMain/kotlin/com/attafitamim/kabin/core/database/Migration.kt
@@ -3,8 +3,8 @@ package com.attafitamim.kabin.core.database
 import app.cash.sqldelight.db.SqlDriver
 
 abstract class Migration(
-    val startVersion: Int,
-    val endVersion: Int
+    val startVersion: Long,
+    val endVersion: Long
 ) {
 
     abstract suspend fun migrate(sqlDriver: SqlDriver)
